repo integration, screenshot modal and bh api
Integrated the use of the repository component for class instances to be
shared easily and added a view screenshot modal for viewing screenshots
alternatively. You can still middle click the view screenshot link to
open in a full tab.

Additions to the BugHerd Api wrapper.

+++ src/main/ui/console.js
@@ -502,10 +502,12 @@ define(
                 );
                 
                 // console destructor tool
-                this.createTool("destroy").on(
-                    "click",
-                    self.destroy
-                );
+                if (config.gui.console.allowDestruction) {
+                    this.createTool("destroy").on(
+                        "click",
+                        self.destroy
+                    );
+                }
                 
                 // clear tool
                 this.createTool("clear").on(
